PFS: use libfsdriver
In order to avoid creating libfsdriver exceptions, two changes to VFS
are necessary:

- the returned position field for reads/writes is no longer abused to
  return the new pipe size; VFS is perfectly capable of updating the
  size itself;
- during system startup, PFS is now sent a mount request, just like all
  other file systems.

In proper "two steps forward, one step back" fashion, the latter point
has the consequence that PFS can no longer drop its privileges at
startup.  This is probably best resolved with a more general solution
for all boot image system services.  The upside is that PFS no longer
needs to be linked with libc.

+++ minix/fs/pfs/inode.h
@@ -11,12 +11,11 @@ EXTERN struct inode {
   nlink_t i_nlinks;		/* how many links to this file */
   uid_t i_uid;			/* user id of the file's owner */
   gid_t i_gid;			/* group number */
-  off_t i_size;			/* current file size in bytes */
+  size_t i_size;			/* current file size in bytes */
   time_t i_atime;		/* time of last access (V2 only) */
   time_t i_mtime;		/* when was file data last changed */
   time_t i_ctime;		/* when was inode itself changed (V2 only)*/
 
-  /* The following items are not present on the disk. */
   dev_t i_dev;			/* which device is the inode on */
   dev_t i_rdev;			/* which special device is the inode on */
   ino_t i_num;			/* inode number on its (minor) device */
@@ -34,6 +33,5 @@ EXTERN TAILQ_HEAD(unused_inodes_t, inode)  unused_inodes;
 
 /* inode hashtable */
 EXTERN LIST_HEAD(inodelist, inode)         hash_inodes[INODE_HASH_SIZE];
-
 
 #endif
